Move method to be test suite only
